\name{organisms}
\docType{data}
\alias{organisms}
\alias{groupTypes}
\alias{FEA_tools}
\alias{GOEvidenceCodes}
\title{FGNet data}
\description{
FGNet package includes the following data files: organisms, groupTypes and FEA_tools
}
\value{
\itemize{
\item{organisms: Matrix with the supported organisms' name and package.}

\item{groupTypes: Matrix with the group types supported by FGNet (cluster, metagroup and gene-term set).}

\item{FEA_tools:  Matrix with info about the FEA tools supported by FGNet.}

\item{GOEvidenceCodes:  Matrix with info about GO evidence codes. (from \url{http://geneontology.org/page/guide-go-evidence-codes}, on nov. 26, 2014.)}

}
}

\usage{
    data(organisms)
    organisms
    
    data(groupTypes)
    groupTypes
    
    data(FEA_tools)
    FEA_tools
    
    data(GOEvidenceCodes)
    GOEvidenceCodes
}

